% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{AS204}
\alias{AS204}
\title{AS204 algorithm}
\usage{
AS204(
  c,
  lambda,
  mult = rep(1, length(lambda)),
  delta = rep(0, length(lambda)),
  maxit = 1e+05,
  eps = 1e-10,
  mode = 1
)
}
\description{
AS204 algorithm
}
\author{
Diego Garrido-Martín
}
\keyword{internal}
