% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{biomarkers}
\alias{biomarkers}
\title{Simulated Measurements of Five Disease Biomarkers}
\format{
A matrix with 100 rows and 5 numerical variables:

\describe{
  \item{biomarker1}{levels of biomarker1}
  \item{biomarker2}{levels of biomarker2}
  \item{...}{}
}
}
\usage{
data(biomarkers)
}
\description{
A simulated dataset containing the levels of 5 biomarkers, 
measured in 100 individuals, with different scales. 
Missing observations appear as \code{NA}.
}
\author{
Diego Garrido-Martín
}
\keyword{datasets}
