% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm2.R
\name{mlm2}
\alias{mlm2}
\title{Multivariate linear models}
\usage{
mlm2(formula, data, distance = "euclidean", contrasts = NULL, ...)
}
\arguments{
\item{formula}{object of class "\code{\link{formula}}": a symbolic 
description of the model to be fitted. The LHS can be either a 
\code{\link{matrix}} or a \code{\link{data.frame}} with the response 
variables, a distance matrix or a distance object of class \code{\link{dist}}.
Note the distance should be euclidean.}

\item{data}{an optional data frame, list or environment (or object 
coercible by \code{\link{as.data.frame}} to a data frame) containing the 
variables in the model. If not found in data, the variables are taken from 
\code{environment(formula)}, typically the environment from which \code{mlm}
is called.}

\item{distance}{data transformation if the formula LHS is not a distance matrix. 
One of c("\code{euclidean}", "\code{hellinger}"). Default is "\code{euclidean}".}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of 
\code{\link{model.matrix.default}}. Default is "\code{\link{contr.sum}}" 
for ordered factors and "\code{\link{contr.poly}}" for unordered factors. 
Note that this is different from the default setting in \code{\link{options}
("contrasts")}.}

\item{...}{additional arguments to be passed to \code{\link{lm}}.}
}
\value{
\code{mlm} returns an object of \code{\link{class}} c("mlm2", "mlm", "lm"); a list containing:
\item{coefficients}{a named vector of coefficients.}
\item{residuals}{the residuals, that is response minus fitted values.}
\item{fitted.values}{the fitted mean values.}
\item{rank}{the numeric rank of the fitted linear model.}
\item{df.residual}{the residual degrees of freedom.}
\item{call}{the matched call.}
\item{terms}{the terms object used (transformed response).}
\item{contrasts}{(only where relevant) the contrasts used.}
\item{xlevels}{(only where relevant) a record of the levels of the factors used in fitting.}
\item{model}{the model frame used (transformed response).}
\item{na.action}{(where relevant) information returned by \code{\link{model.frame}} on the special handling of NAs.}
\item{transformation}{transformations of the response matrix.}
In addition, non-null fits will have components assign, effects and qr 
relating to the linear fit. See \code{\link{lm}} for further details. 
Note the multivariate fit is done on the transformed response \code{"Y"}.
}
\description{
Transforms the response variables and fits a multivariate
linear model using \code{\link{lm}}.
}
\details{
A "\code{Y}" matrix is obtained after projecting into euclidean space 
(as in multidimensional scaling) and centering the original response 
variables. Then, the multivariate fit obtained by \code{\link{lm}} can be 
used to obtain sums of squares (I, II, III), pseudo F statistics and 
asymptotic p-values for the explanatory variables in a non-parametric manner.
}
\seealso{
\code{\link{summary.mlm2}}
}
\author{
Diego Garrido-Martín
}
