% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm2.R
\name{mlmdist}
\alias{mlmdist}
\title{Distance matrix computation for Euclidean family distances}
\usage{
mlmdist(X, method = "euclidean")
}
\arguments{
\item{d}{distance matrix}

\item{distance}{distance to be applied. One of c("euclidean", "hellinger")}

\item{tol}{eigenvalues with absolute value <= tol are considered 0.}
}
\description{
This function computes and returns the distance matrix computed by using 
the specified distance measure to compute the distances between the rows 
of a data matrix.
}
\details{
Available distance measures are (written for two vectors x and y):
\code{euclidean}:

\code{hellinger}:
}
