% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{mlmproject}
\alias{mlmproject}
\title{Project into euclidean space}
\usage{
mlmproject(dmat, k = NULL, tol = 1e-12)
}
\arguments{
\item{dmat}{distance matrix as obtained by \code{\link{dist}} or 
\code{\link{mlmdist}}.}

\item{k}{number of columns of the original matrix. Default is \code{NULL}.}

\item{tol}{values below this threshold will be considered 0}
}
\description{
This function obtains the projection in an euclidean space of the original
variables using its distance matrix, in a similar way to multidimensional
scaling.
}
\details{
When \code{k} is provided, \code{\link{eigs_sym}} function is used, 
instead of \code{\link{eigen}}, to compute only the top \code{k} eigenvalues 
and the corresponding eigenvectors.
}
