% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\docType{data}
\name{patients}
\alias{patients}
\title{Patients}
\format{
A matrix with 100 rows and 3 variables:
\describe{
  \item{gender}{Gender of the patient (factor with levels: \code{male} and 
  \code{female})}
  \item{age}{Age of the patient (numerical)}
  \item{disease}{Disease status of the patient (ordered factor with levels
  \code{healthy}, \code{"mild"}, \code{"severe"})}
}
}
\usage{
patients
}
\description{
A simulated dataset containing the gender, age and disease status of 100
individuals. Missing observations appear as \code{NA}.
}
\keyword{datasets}
