% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{pv.f}
\alias{pv.f}
\title{Compute asymptotic P-values}
\usage{
pv.f(f, lambda, df.i, df.e, acc = 1e-14)
}
\arguments{
\item{f}{pseudo-F statistic.}

\item{lambda}{eigenvalues}

\item{df.i}{degrees of freedom of the variable}

\item{df.e}{residual degrees of freedom}

\item{acc}{precision limit}
}
\description{
Description
}
\details{
Details
}
