% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{pv.ss}
\alias{pv.ss}
\title{Compute asymptotic P-values}
\usage{
pv.ss(ss, lambda, df.i, eps = 1e-14)
}
\arguments{
\item{ss}{numerator of the pseudo-F statistic.}

\item{lambda}{eigenvalues.}

\item{df.i}{degrees of freedom of the numerator of the test statistic.}

\item{eps}{precision limit.}
}
\description{
Description
}
\details{
Details
}
\author{
Diego Garrido-Martín
}
