% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm2.R
\name{summary.mlm2}
\alias{summary.mlm2}
\title{Summary method for Non-parametric Multivariate Analysis of Variance}
\usage{
\method{summary}{mlm2}(object, type = "II", ...)
}
\arguments{
\item{object}{an object of class "\code{mlm2}"}

\item{type}{type of sum of squares (either "I", "II", "III" or 1, 2, 3).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A summary method for class "\code{mlm2}".
}
