% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{mlm}
\alias{mlm}
\title{Multivariate linear models}
\usage{
mlm(formula, data, transform = NULL, type = "II", contrasts = NULL, ...)
}
\arguments{
\item{formula}{object of class "\code{\link{formula}}": a symbolic 
description of the model to be fitted.}

\item{data}{an optional data frame containing the variables in the model. 
If not found in data, the variables are taken from \code{environment(formula)}, 
typically the environment from which \code{mlm} is called.}

\item{transform}{transformation of the response variables: \code{NULL}, 
"\code{sqrt}" or "\code{log}". Default is \code{NULL}.}

\item{type}{type of sum of squares: "\code{I}", "\code{II}" or "\code{III}". 
Default is "\code{II}".}

\item{contrasts}{an optional list. See \code{contrasts.arg} in 
\code{\link{model.matrix.default}}. Default is "\code{\link{contr.sum}}" 
for ordered factors and "\code{\link{contr.poly}}" for unordered factors. 
Note that this is different from the default setting in 
\code{\link{options}("contrasts")}.}

\item{...}{additional arguments to be passed to other functions.}
}
\value{
\code{mlm} returns an object of \code{\link{class}} "MLM", a list containing:
\item{call}{the matched call.}
\item{aov.tab}{ANOVA table with Df, Sum Sq, Mean Sq, F values, partial R2 and P values.}
\item{type}{the type of sum of squares (\code{I}, \code{II} or \code{III}).}
\item{precision}{the precision in P value computation.}
\item{transform}{the transformation applied to the response variables.}
\item{fit}{the multivariate fit done on the transformed and centered response variables.}
}
\description{
Fits a multivariate linear model and computes test statistics and p-values
for a set of predictors in a non-parametric manner.
}
\details{
A \code{Y} matrix is obtained after transforming and centering the original 
response variables. Then, the multivariate fit obtained by \code{\link{lm}} 
can be used to compute sums of squares (\code{I}, \code{II} or \code{III}), 
pseudo F statistics and asymptotic p-values for the explanatory 
variables in a non-parametric manner.
}
\seealso{
\code{\link{lm}}
}
\author{
Diego Garrido-Martín
}
