% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{mlmdist}
\alias{mlmdist}
\title{Distance matrix computation for Euclidean distances}
\usage{
mlmdist(X, method = "euclidean")
}
\arguments{
\item{X}{distance matrix}

\item{method}{distance to be applied. One of c("euclidean", "hellinger")}
}
\description{
This function computes and returns the distance matrix obtained by using 
the specified distance measure to compute the distances between the rows 
of a data matrix.
}
\details{
Available distance measures are (written for two vectors x and y):
\code{euclidean}:

\code{hellinger}:
}
