% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{mlmtst}
\alias{mlmtst}
\title{Compute test statistic}
\usage{
mlmtst(fit, type = "II")
}
\arguments{
\item{fit}{multivariate fit obtained by \code{\link{lm}}.}

\item{type}{type of sum of squares (\code{I}, \code{II} or \code{III}). Default is \code{II}.}
}
\description{
This function computes the degrees of freedom, sum of squares, partial R2 and
pseudo F statistic for each explanatory variable from \code{fit}.
}
\details{
Different types of sums of squares are available.
}
