% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{patients}
\alias{patients}
\title{Simulated Metadata for 100 Patients}
\format{
A matrix with 100 rows and 3 variables:

\describe{
  \item{age}{Age of the patient (numerical)}
  \item{gender}{Gender of the patient (factor with levels: "male" and 
  "female")}
  \item{status}{Disease status of the patient (ordered factor with levels:
  "healthy", "mild" and "severe")}
}
}
\usage{
data(patients)
}
\description{
A simulated dataset containing the age, gender and disease status of 100
individuals. Missing observations appear as \code{NA}.
}
\author{
Diego Garrido-Martín
}
\keyword{datasets}
