% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manta.R
\name{printCoefmat.mp}
\alias{printCoefmat.mp}
\title{Print Coefficient Matrices (Multiple P-value Precision Limits)}
\usage{
printCoefmat.mp(
  x,
  digits = max(3L, getOption("digits") - 2L),
  signif.stars = getOption("show.signif.stars"),
  signif.legend = signif.stars,
  dig.tst = max(1L, min(5L, digits - 1L)),
  cs.ind = 1:k,
  tst.ind = k + 1,
  zap.ind = integer(),
  P.values = NULL,
  has.Pvalue = nc >= 4 && substr(colnames(x)[nc], 1, 3) == "Pr(",
  eps.Pvalue = .Machine$double.eps,
  na.print = "NA",
  ...
)
}
\description{
Modification of \code{\link{printCoefmat}} to use multiple P-value 
precision limits.
}
\seealso{
\code{\link{printCoefmat}}
}
\author{
Diego Garrido-Martín
}
\keyword{internal}
