% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{pv.ss}
\alias{pv.ss}
\title{Compute asymptotic P-values}
\usage{
pv.ss(ss, lambda, df, eps = 1e-14, eps.updt = 2)
}
\arguments{
\item{ss}{numerator of the pseudo-F statistic.}

\item{lambda}{eigenvalues.}

\item{df}{degrees of freedom of the numerator of the test statistic.}

\item{eps}{precision limit.}

\item{eps.updt}{factor by which precision limit is updated.}
}
\description{
Compute asymptotic P-values
}
\author{
Diego Garrido-Martín
}
\keyword{internal}
