% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{pv.ss}
\alias{pv.ss}
\title{Compute asymptotic P-values}
\usage{
pv.ss(f, lambda, df.i, eps = 1e-14, tol = 0.001)
}
\arguments{
\item{lambda}{eigenvalues.}

\item{df.i}{degrees of freedom of the numerator of the test statistic.}

\item{eps}{precision limit.}

\item{tol}{lambda/sum(lambda) > tol.}

\item{ss}{numerator of the pseudo-F statistic.}
}
\description{
Description
}
\details{
Details
}
